pyproject.toml
[tool.poetry]
name = "timesweeps-pump"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include="timesweeps_pump"}]

[tool.poetry.dependencies]
python = "^3.10"
python-syringe-pump = "^0.2.0"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
jupyter = "^1.0.0"
black = "^23.9.1"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
ignore = ["F704", "E510"]

[tool.isort]
profile = "black"
